requirements.txt
slack_bolt==1.21.3
python-dotenv==1.0.1
flask==3.0.0
